+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2024-07-15
+
+- Move from `requests` to `aiohttp`
+- Make library async
+- Updated examples
+
 ## [1.0.0] - 2024-07-15
 
 - Initial release
 
-[unreleased]: https://github.com/itsnebulalol/pyrd/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/itsnebulalol/pyrd/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/itsnebulalol/pyrd/releases/tag/v2.0.0
 [1.0.0]: https://github.com/itsnebulalol/pyrd/releases/tag/v1.0.0